+++ src/compute-engine/collection-utils.ts
@@ -1346,6 +1346,7 @@ export function defaultCollectionHandlers(
   };
   if (def.isCollection) result.isCollection = def.isCollection;
   if (def.isLazy) result.isLazy = def.isLazy;
+  if (def.elementMemo) result.elementMemo = def.elementMemo;
   if (def.eltsgn) result.eltsgn = def.eltsgn;
   if (def.elttype) result.elttype = def.elttype;
   if (def.at) {